Factor out a call::buffers module
